+++ consensus/merge/src/main/java/org/hyperledger/besu/consensus/merge/blockcreation/MergeCoordinator.java
@@ -81,13 +81,14 @@ public class MergeCoordinator implements MergeMiningCoordinator, BadChainListene
   private static final double TRY_FILL_BLOCK = 1.0;
 
   private static final long DEFAULT_TARGET_GAS_LIMIT = 36_000_000L;
-  // testnets might have higher gas limits than mainnet
+  // testnets might have higher gas limits than mainnet and are incrementally updated
   private static final long DEFAULT_TARGET_GAS_LIMIT_TESTNET = 36_000_000L;
+  private static final long NEXT_STEP_TARGET_GAS_LIMIT_TESTNET = 60_000_000L;
 
   private static final Map<BigInteger, Long> TESTNET_CHAIN_IDS =
       Map.of(
-          BigInteger.valueOf(11155111), 60_000_000L, // Sepolia
-          BigInteger.valueOf(17000), DEFAULT_TARGET_GAS_LIMIT_TESTNET, // Holesky
+          BigInteger.valueOf(11155111), NEXT_STEP_TARGET_GAS_LIMIT_TESTNET, // Sepolia
+          BigInteger.valueOf(17000), NEXT_STEP_TARGET_GAS_LIMIT_TESTNET, // Holesky
           BigInteger.valueOf(560048), DEFAULT_TARGET_GAS_LIMIT_TESTNET, // Hoodi
           BigInteger.valueOf(39438135), DEFAULT_TARGET_GAS_LIMIT_TESTNET // Ephemery
           );